# dyno: changes to improve `--dyno-scope-bundled` performance