--- a/test/separate_compilation/serialization/buildAll.compopts
+++ b/test/separate_compilation/serialization/buildAll.compopts
@@ -1,1 +1,1 @@
-testGenLib.dyno MyMod.dyno OtherMod.dyno chpl_standard.dyno --main-module testGenLib --dyno-debug-trace
+testGenLib.dyno MyMod.dyno OtherMod.dyno chpl_standard.dyno --main-module testGenLib --dyno-debug-print-parsed-files