--- a/test/separate_compilation/serialization/buildAll.good
+++ b/test/separate_compilation/serialization/buildAll.good
@@ -1,3 +1,3 @@
-2 { fileTextQuery ("buildAll.chpl") 
+Parsed file: buildAll.chpl
 MYPREFIX: doMyM-doMyM-doMyM
 MYPREFIX: doMrehtO-doMrehtO-doMrehtO