--- a/test/separate_compilation/serialization/testGenLib.prediff
+++ b/test/separate_compilation/serialization/testGenLib.prediff
@@ -1,10 +1,7 @@
 #!/bin/bash
 
-grep "fileTextQuery (\|^MYPREFIX" $2 > $2.tmp
+grep "Parsed file:\|^MYPREFIX" $2 > $2.tmp
 
-cat $2.tmp | sed 's/hash: .*$//g' > $2
-
-mv $2 $2.tmp
 cat $2.tmp | sed "s|${CHPL_HOME}|\$CHPL_HOME|g" > $2
 
 mv $2 $2.tmp