… test/separate_compilation/serialization/testGenLib.MyMod.good
… test/separate_compilation/serialization/testGenLib.OtherMod.good
… test/separate_compilation/serialization/testGenLib.both.good